Moved property comments into @property jsdocs, added inventory event dtos, seperated order dtos into seperate files, made jsdocs equal to shipment definitions
